+++ CHANGELOG.md
@@ -18,6 +18,10 @@ If you were at `firehose-core` version `1.0.0` and are bumping to `1.1.0`, you s
 
 * Fix egress bytes calculation when running in noop or dev mode with specified output debug modules.
 
+### Firehose
+
+* Added `firehose_session_denied_counter` with `reason` label, increment each time a session is refused with the reason why it was refused.
+
 ## v1.12.1
 
 ### Substreams